requirements.txt
Flask==3.0.0

Flask-CORS==4.0.0

requests==2.31.0

beautifulsoup4==4.12.2

python-dotenv==1.0.0

pymongo==4.6.0

dnspython==2.4.2

reportlab==4.0.7

cryptography==41.0.7

urllib3==2.1.0

validators==0.22.0

report_generator ==0.1.0